% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{allIsInteger}
\alias{allIsInteger}
\title{check whether vector represents all integer values, not that the same as
\code{is.integer}}
\usage{
allIsInteger(x, tol = 1e-09, na.rm = TRUE)
}
\arguments{
\item{x}{is a vector to be tested}

\item{tol}{single numeric, default if less than 1e-9 from an integer then
considered an integer.}

\item{na.rm}{single logical, passed on to \code{all}}
}
\value{
logical scalar
}
\description{
check whether all the items of input vector are integer as.integer
}
