% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{allIsNumeric}
\alias{allIsNumeric}
\title{check whether character vector represents all numeric values}
\usage{
allIsNumeric(x, extras = c(".", "NA", NA))
}
\arguments{
\item{x}{is a character vector to be tested}

\item{extras}{is a vector of character strings which are counted as NA
values, defaults to '.' and 'NA'. Also allow \code{NA}.}
}
\value{
logical scalar
}
\description{
check whether all the items of input vector are numeric without throwing
warning derived from Hmsic package
}

