% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{areNumeric}
\alias{areNumeric}
\title{which elements of a vector are numeric}
\usage{
areNumeric(x, extras = c(".", "NA", NA))
}
\arguments{
\item{x}{vector}
}
\value{
logical vector of same length as input
}
\description{
test without throwing a warning
}
\examples{
areNumeric(c("1","2","3"))
areNumeric(c("1L", "2.2"))
areNumeric(c("NA", NA, ".", "", "-1.9"))
}

