% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{asCharacterNoWarn}
\alias{asCharacterNoWarn}
\title{convert factor or vector to character without warnings}
\usage{
asCharacterNoWarn(x)
}
\arguments{
\item{x}{is a vector, probably of numbers of characters}
}
\value{
character vector, may have NA values
}
\description{
correctly converts factors to vectors, and then converts to
  character, which may silently introduce NAs
}

