% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.r
\docType{data}
\name{bad_input}
\alias{bad_input}
\title{bad input data for tests}
\format{\preformatted{List of 41
 $     : NULL
 $     : logi NA
 $     : list()
 $     :List of 2
  ..$ : num 1
  ..$ : num 2
 $ jack: chr "test"
 $     : num [1:3] 1 2 3
 $     :'data.frame':	0 obs. of  0 variables
 $     : logi NA
 $ this: chr "jack"
 $ j   : int [1:10] 1 2 3 4 5 6 7 8 9 10
 $ k   : int [1:10] 11 12 13 14 15 16 17 18 19 20
 $     : logi NA
 $     : logi NA
 $     : logi NA
 $     : logi NA
 $     : num 1.42e+09
 $     : num 78796800
 $     : num 94694400
 $     : num 1.26e+08
 $     : num 1.58e+08
 $     : num 1.89e+08
 $     : num 2.21e+08
 $     : num 2.52e+08
 $     : num 2.84e+08
 $     : num 3.16e+08
 $     : num 3.63e+08
 $     : num 3.94e+08
 $     : num 4.26e+08
 $     : num 4.89e+08
 $     : num 5.68e+08
 $     : num 6.31e+08
 $     : num 6.63e+08
 $     : num 7.1e+08
 $     : num 7.41e+08
 $     : num 7.73e+08
 $     : num 8.2e+08
 $     : num 8.68e+08
 $     : num 9.15e+08
 $     : num 1.14e+09
 $     : num 1.23e+09
 $     : num 1.34e+09
}}
\usage{
bad_input
}
\description{
a variety of horrible data
}
\keyword{sysdata}

