% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/df.r
\name{binaryCols}
\alias{binaryCols}
\alias{nonBinaryCols}
\title{return names of fields which consist of only 0 and 1}
\usage{
binaryCols(x)

nonBinaryCols(x)
}
\arguments{
\item{x}{data frame}
}
\value{
vector of column names
}
\description{
assumes columns are numeric, doesn't think about factors
#TODO check for factors, etc.
}

