% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{build_formula}
\alias{build_formula}
\alias{buildLinearFormula}
\title{build simple linear formula from variable names}
\usage{
build_formula(left, right)

buildLinearFormula(left, right)
}
\arguments{
\item{left}{character vector}

\item{right}{character vector}
}
\value{
formula
}
\description{
build simple linear formula from variable names given by two
  character vectors. TODO: allow unquoted names.
}
\examples{
print(f <- build_formula(left = "A", right = c("B", "C")))
class(f)
build_formula(left = "Species", right = names(iris)[1:4])
}
