% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{assignCache}
\alias{assignCache}
\alias{assignCacheByFun}
\alias{cache}
\title{cache and assign data to an environment}
\usage{
assignCache(value, var, from = NULL, to = NULL, ...,
  envir = parent.frame(), verbose = FALSE)

cache(value, var, from = NULL, to = NULL, cacheDir = NULL,
  envir = parent.frame(), force = FALSE, assign = FALSE,
  format = "\%Y-\%m-\%d", verbose = FALSE)

assignCacheByFun(fun, var, from = NULL, to = NULL, ..., cacheDir = NULL,
  envir = parent.frame())
}
\arguments{
\item{value}{value to be lazy evaluated. Consider also enabling a call, or
expression. \code{eval} could be called on it, so a constructed expression
can be passed in. Since scoping changes, this should be self contained.}

\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{from}{This should be a single YYYY-MM-DD string. There are no
absolute requirements about the format, but it is intended to be the same
format as used in a SQL query.}

\item{to}{see start date}

\item{...}{arguments passed on to other functions}

\item{envir}{environment to assign to, defaults to the calling frame}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}

\item{force}{single logical value whether to force re-writing of cache if
data already exists}

\item{assign}{single logical value whether to assign the cached data to an
environment}

\item{format}{single character value with date format}

\item{fun}{function which accepts from and to arguments used by
\code{assignCacheByFun}}
}
\value{
invisibly returns the value assigned
}
\description{
Same as assign, but will only touch cache if the variable isn't
  already loaded.
}
\seealso{
Other cache: \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveInDataDir}}, \code{\link{saveToCache}}
}

