% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{combn_subset}
\alias{combn_subset}
\title{all unique combinations of a vector and all its non-zero subsets}
\usage{
combn_subset(x)
}
\arguments{
\item{x}{vector to be subsetted and combined}
}
\value{
list of vectors with all combinations of x and its subsets
}
\description{
all unique combinations of a vector and all its non-zero subsets
}
\examples{
combn_subset(c("a", "b"))
combn_subset(c(10, 20, 30))
combn_subset(NULL)
}
