% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{countNonNaCumulative}
\alias{countNonNaCumulative}
\title{running totals of number of non-NA values in consecutive fields}
\usage{
countNonNaCumulative(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
vector of cumulative non-NA counts with names corresponding to the
  given data frame
}
\description{
counts non-NA fields in first field, then progreses through
  fields, OR new field and saves running total for each field TODO: tests
}
