% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{countNonNaPairs}
\alias{countNonNaPairs}
\title{count which combinations of fields have at least one non-NA}
\usage{
countNonNaPairs(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
matrix with nrow and ncol being the number of fields in the given
  dataframe
}
\description{
cycles through the given data frame twice, and applies logical
  OR to all elements of each column it then counts how many of these pairs
  are not-na, i.e. have at least one non-NA value TODO: tests
}

