% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{countNotNumeric}
\alias{countNotNumeric}
\title{count non-numeric elements}
\usage{
countNotNumeric(x)
}
\arguments{
\item{x}{is usually a charcter vector}
}
\value{
integer
}
\description{
counts the number of non-numeric elements in a vector, without
  throwing warnings
}
\details{
did have \code{extras = c(".", "NA"))}
}
