% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{countNumeric}
\alias{countNumeric}
\title{count numeric elements}
\usage{
countNumeric(x)
}
\arguments{
\item{x}{is usually a character vector}
}
\value{
integer
}
\description{
counts the number of numeric elements in a vector, without
  throwing warnings
}

