% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dput_expect_that.r
\name{dput_expect_equal}
\alias{dput_expect_equal}
\title{dput a testthat test}
\usage{
dput_expect_equal(...)
}
\arguments{
\item{...}{expressions}
}
\value{
character vector with each element containing an R expression with
  \code{expect_equal} test case corresponding to the evaluated input
  expressions.
}
\description{
Generate an R expression containing a \code{testthat} expectation for the
given expression and its result. This is useful when you know that a certain
output is correct, and wish to generate a test case to reflect this.
}
\examples{
dput_expect_equal("a" \%nin\% c("b", "c", "d"))
}
