% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{dropDuplicateFields}
\alias{dropDuplicateFields}
\title{drop duplicate fields}
\usage{
dropDuplicateFields(df, verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}
}
\value{
data frame without duplicate fields
}
\description{
compares all data in each field to every other field, and drops
  the latter match. Will find multiple matches. Doesn't do any type
  conversions yet. This is purely by content, not by field name.
}

