% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.r
\name{\%eine\%}
\alias{\%eine\%}
\alias{\%ine\%}
\title{\%in\%/match equivalent for two \code{environment} arguments}
\usage{
x \%eine\% table

x \%ine\% table
}
\description{
\code{x} and \code{table} use the same as with
  \code{base::match}. Lookup is done based on environment element names;
  contents are ignored. For \code{\%ine\%}, \code{x} is a vector, and
  \code{table} is the environment to search. For \code{\%eine\%}, both
  arguments are environments.
}
\section{Functions}{
\itemize{
\item \code{\%ine\%}: search vector of values in an environment
}}

\keyword{internal}
