% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.r
\docType{data}
\name{extreme_numbers}
\alias{extreme_numbers}
\title{extreme numbers}
\format{\preformatted{ num [1:6] 2.15e+09 -2.15e+09 2.23e-308 1.80e+308 -2.23e-308 ...
}}
\usage{
extreme_numbers
}
\description{
very biggest and smallest non-zero numbers the current machine
  can handle, positive and negative.
}
\keyword{sysdata}

