% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{factorToDataframeLogical}
\alias{factorToDataframeLogical}
\title{convert factor into a data.frame of logicals}
\usage{
factorToDataframeLogical(fctr, prefix = deparse(substitute(fctr)), sep = "",
  na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{fctr}{factor}

\item{prefix}{defaults to "f" to pre-pend the factor level when constructing
the data frame columns names}

\item{sep}{scalar character, introduced between factor names and levels when
forming new data frame column names}

\item{na.rm}{logical scalar: if NA data and/or NA levels, then covert to NA
strings and expand these as for any other factor}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}
}
\value{
data.frame with columns of logicals
}
\description{
converts a single factor into a data.frame with multiple T/F
  fields, one for each factor
}

