% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{fillMissingCombs}
\alias{fillMissingCombs}
\title{fill out missing combinations of factors with NA}
\usage{
fillMissingCombs(df)
}
\arguments{
\item{df}{data frame}
}
\details{
Adapated from http://www.cookbook-r.com/Manipulating_data/\
Summarizing_data/#using-aggregate
}

