% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_better.r
\name{filter_better}
\alias{filter_better}
\title{filter data with diagnostics}
\usage{
filter_better(x, expr, verbose = TRUE)
}
\arguments{
\item{x}{data frame}

\item{expr}{expression in the context of the data frame, i.e. the terms
should be column names.}

\item{verbose}{logical default is TRUE}
}
\value{
filtered data frame
}
\description{
applies an expression to a data frame, and gives information
  about the numbers of dropped rows.
}
