% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{findCacheDir}
\alias{findCacheDir}
\title{find path to the cache directory}
\usage{
findCacheDir(cacheDir = NULL, cacheDirName = "jwcache", verbose = FALSE)
}
\arguments{
\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}

\item{cacheDirName}{single character string, defaults to 'jwcache'. 'cache'
alone is not distinctive, and conflicts with other things, such as the
cache directory in the vignettes directory.}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}
}
\description{
Searches a few likely places: - directory specified in
  \code{cacheDir} - session-wide R option "cacheDir" - \code{cache} directory
  within the working directory - \code{cache} directory within the parent
  directory Fail with error if we have still not found it. Walking up the
  directory tree is both time consuming, low yield, and unreliable, so this
  is no longer done. Instead, a fall-back cache directory is used based on an
  /code{options} setting jwutil.fallbackCacheDir, which defaults to ~/jwutil.
  Would prefer to use /tmp but this doesn't exist on Windows, and using the R
  built-in to generate a temporary directory is not possible with a static
  option. Generating it on-the-fly would need passing the cache dir around
  between functions or creating a singleton, all of which get too complicated
  for this corner case. Of note, R CMD check creates a distinct directory
  tree with all the testing files, and may not include the \code{cache}
  directory from the working tree, e.g. if the cache directory is in
  \code{.Rbuildignore}.
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveInDataDir}}, \code{\link{saveToCache}}
}

