% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{findCacheFilePath}
\alias{findCacheFilePath}
\title{find path to a file in cache directory}
\usage{
findCacheFilePath(filename, cacheDir = NULL)
}
\arguments{
\item{filename}{character vector of file names}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveInDataDir}}, \code{\link{saveToCache}}
}

