% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{flattenList}
\alias{flattenList}
\title{flatten a list}
\source{
https://stackoverflow.com/questions/8139677/\
  how-to-flatten-a-list-to-a-list-without-coercion
}
\usage{
flattenList(..., na.rm = FALSE)
}
\arguments{
\item{...}{list or any set of objects which will be made into a list, may
include lists and nested lists}

\item{na.rm}{will drop NA values if TRUE}
}
\value{
list without nested lists, objects with preserved data types
}
\description{
unlike unlist, this function returns a list of objects of
  different data types, but removes any depth
}

