% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factor_fields.r
\name{get_factor_fields}
\alias{get_factor_fields}
\alias{getFactorNames}
\title{get names of the factor fields in a data frame}
\usage{
get_factor_fields(x, consider = names(x))

getFactorNames(x, consider = names(x))
}
\arguments{
\item{x}{data frame}

\item{consider}{character vector of field names of the data frame to test,
default is to use all of them.}
}
\value{
vector
}
\description{
Get the names of those fields in a data frame which are factors.
}
