% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.r
\name{get_na_fields}
\alias{get_na_fields}
\alias{getNAFields}
\alias{get_non_na_fields}
\alias{getNonNAFields}
\title{get NA field names from data frame}
\usage{
get_na_fields(dframe, na_ish = FALSE, extra_na = NULL)

getNAFields(dframe, na_ish = FALSE, extra_na = NULL)

get_non_na_fields(dframe)

getNonNAFields(dframe)
}
\arguments{
\item{dframe}{data.frame}

\item{na_ish}{Logical, if `TRUE`, also consider NA-like strings, using
`is_na_ish`}

\item{extra_na}{passed on to `is_na_ish`}
}
\value{
vector of names of fields which contain any NA values, length zero if
  no matches
}
\description{
Get the names of any columns in a data frame which have NA
  values.
}
\section{Functions}{
\itemize{
\item \code{getNAFields}: Deprecated

\item \code{getNonNAFields}: Deprecated
}}

