% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric_fields.r
\name{get_numeric_char_field_names}
\alias{get_numeric_char_field_names}
\alias{get_numeric_field_names}
\alias{get_numeric_fields}
\title{Find columns which are numeric}
\usage{
get_numeric_char_field_names(x, invert = FALSE, attrition = 0.05)

get_numeric_field_names(x, invert = FALSE)

get_numeric_fields(x, invert = FALSE)
}
\arguments{
\item{x}{Data frame}

\item{invert}{Logical, if \code{FALSE} -- the default -- the numeric fields
are returned, otherwise, non-numeric fields are returned.}

\item{attrition}{If less than this proportion of rows become \code{NA} on
conversion to numeric, then accept this is a numeric column after all.}
}
\description{
Get field names, or the data itself, of fields in a data frame which are
numeric, or numeric-like characters.
}
