% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{getDropped}
\alias{getDropped}
\title{get items or numerics that would be dropped in a merge}
\usage{
getDropped(x, y)
}
\arguments{
\item{x}{vector or factor}

\item{y}{vector or factor}
}
\value{
list of two vectors
}
\description{
converts both vectors to numeric. This simulates merging when
  one key is character (but contains integer numbers), and another key is
  stored as integer.
}

