% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{getFromCache}
\alias{getFromCache}
\alias{loadFromCache}
\title{get or load a variable from cache}
\usage{
getFromCache(var, from = NULL, to = NULL, cacheDir = NULL,
  verbose = FALSE)

loadFromCache(var, ..., envir = parent.frame(), inherits = FALSE)
}
\arguments{
\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{from}{This should be a single YYYY-MM-DD string. There are no
absolute requirements about the format, but it is intended to be the same
format as used in a SQL query.}

\item{to}{see start date}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}

\item{...}{params to pass on to getFromCache}

\item{envir}{environment in which to load the variable, defaults to the
calling frame.}

\item{inherits}{single logical, passed on to \code{assign}}
}
\description{
Get data from the cache, optionally with date ranges. load from
  cache also returns the variable, but additionally saves the data by its
  undated cache name in the given environment, defaulting to the calling
  frame.
}

