% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{invwhich}
\alias{invwhich}
\title{inverse which}
\usage{
invwhich(which, len = max(which))
}
\arguments{
\item{which}{integer vector of indices, as would be produced by \code{which}}

\item{len}{integer scalar: length of return vector, defaults to
\code{max(which)}}
}
\value{
logical vector of length \code{length}
}
\description{
for a given vector of ordinals which would reference items in a
  vector, list etc, \code{invwhich} returns a logical vector with TRUE for
  the cited positions. If length is not provided, the maximum index is used.
}

