% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.r
\name{is_na_ish}
\alias{is_na_ish}
\title{Determine whether a value is, or should be, `NA`}
\usage{
is_na_ish(x, extra_na = NULL)
}
\arguments{
\item{x}{vector to test}

\item{extra_na}{Additional values to consider equivalent to NA}
}
\description{
Determine whether a value is, or should be, `NA`
}
\examples{
is_na_ish(c(NA, "1"))
is_na_ish(c("NA", "N/A", "NaN"))
is_na_ish(c(NA))
is_na_ish(c(NA))
}
