% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{isCached}
\alias{isCached}
\title{check whether an object is in the cache}
\usage{
isCached(var, from = NULL, to = NULL, cacheDir = NULL)
}
\arguments{
\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{from}{This should be a single YYYY-MM-DD string. There are no
absolute requirements about the format, but it is intended to be the same
format as used in a SQL query.}

\item{to}{see start date}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}
}
\value{
logical, single logical value.

logical TRUE if found in cache
}
\description{
use the same search algorithm as loading from the cache, but
  only report whether the file was there.
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveInDataDir}}, \code{\link{saveToCache}}
}

