% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.r
\name{isFlat}
\alias{isFlat}
\title{determine whether a list is nested}
\usage{
isFlat(x)
}
\arguments{
\item{x}{list}
}
\value{
single logical
}
\description{
determine whether a list is nested
}

