% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isRowSorted}
\alias{isRowSorted}
\title{is every row sorted?}
\usage{
isRowSorted(x)
}
\arguments{
\item{x}{matrix, each row containing ordered or disordered numerics}
}
\description{
Quicky run through rows of a matrix looking for any
non-ascending rows in C++
}
