% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{isValidTime}
\alias{isValidTime}
\title{check if a time is valid in 24h clock}
\usage{
isValidTime(tms, na.rm = FALSE)
}
\arguments{
\item{tms}{is a vector of characters which may represent times}

\item{na.rm}{logical if true, will ignore NA values, otherwise these will
test as invalid.}
}
\value{
logical vector, with NA out if NA given
}
\description{
allow leading and trailing space, optional colon in middle, 2400
  is not allowed. TODO: can lubridate do this better?
}

