% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{jw_df_basics}
\alias{jw_df_basics}
\title{minimal basic pre-processing metrics}
\usage{
jw_df_basics(x, df_list)
}
\arguments{
\item{x}{data.frame input}

\item{df_list}{list of data frames}
}
\description{
minimal basic pre-processing metrics
}

