% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.r
\name{jw_qq_diag}
\alias{jw_qq_diag}
\title{qqplot variables from multiple matching methods}
\usage{
jw_qq_diag(Tr, dat, m_out, qq_cov = names(gm))
}
\description{
qqplot variables from multiple matching methods
}

