% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-build.R
\name{jw_scan_build}
\alias{jw_scan_build}
\title{Build with current Makevars, but with clang scan-build static analysis}
\usage{
jw_scan_build(path = ".", clang = "clang-8",
  scan_build = "scan-build")
}
\arguments{
\item{path}{Path to package root, default is \code{"."}.}

\item{clang}{Path or name of clang compiler executable. Currently `clang-8`
which is what MacOS homebrew currently (early 2019) provides.}

\item{scan_build}{Path or name of scan-build executable. Current
`scan-build`, also from MacOS Homebrew. On linux, this has the LLVM version
suffix, e.g., `scan-build-8`.}
}
\description{
C and C++ compilers are replaced by `scan-build clang`, and restored
afterwards. Other flags and anything else in `~/.R/Makevars` is left alone.
}
