% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jwutil-package.r
\docType{package}
\name{jwutil}
\alias{jwutil}
\alias{jwutil-package}
\title{jwutil.}
\description{
A small set of simple utilities for common data manipulation
  tasks, primarily of data frames. The goal is typically to use built-in
  tools well, without bringing in huge dependencies.
}

