% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{list_named}
\alias{list_named}
\title{Make a list using input argument names as names}
\usage{
list_named(...)
}
\arguments{
\item{...}{arguments whose names become list item names, and whose values
become the values in the list}
}
\description{
Make a list using input argument names as names
}
\examples{
a <- c(1, 2)
b <- c("c", "d")
stopifnot(
  identical(
    named_list(a, b),
    list(a = a, b = b)
  )
)
}
