% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{listTrim}
\alias{listTrim}
\title{trim null or empty values from a list}
\usage{
listTrim(x)
}
\arguments{
\item{x}{list}
}
\value{
trimmed list
}
\description{
delele null/empty entries in a list. Recursively looks through
  list if nested.
}
