% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_to_binary.r
\name{logical_to_binary}
\alias{logical_to_binary}
\alias{logicalToBinary}
\title{Convert logical columns of data frame to 0s and 1s}
\usage{
logical_to_binary(x)

logicalToBinary(x)
}
\arguments{
\item{x}{data frame which may contain logical fields}
}
\value{
data frame without logical fields
}
\description{
Encode TRUE as 1, and FALSE as 0 (integers)
}
\examples{
d <- data.frame(a = c(TRUE, FALSE, TRUE),
                b = c(FALSE, TRUE, FALSE),
                c = c(-1, 0, 1),
                d = c("not", "logical", "values")
                )
logical_to_binary(d)
}
\keyword{manip}
