% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{lsCacheFiles}
\alias{lsCache}
\alias{lsCacheFiles}
\title{list cache contents}
\usage{
lsCacheFiles(pattern = ".*", cacheDir = NULL)

lsCache(pattern = ".*", cacheDir = NULL)
}
\arguments{
\item{pattern}{regex, ".Rdata$" is appended within the function for file name
searches}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{optName}}, \code{\link{renameCache}},
  \code{\link{rmCache}}, \code{\link{saveInDataDir}},
  \code{\link{saveToCache}}
}

