% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.r
\name{lsf}
\alias{lsf}
\title{list all functions in a package}
\usage{
lsf(pkg)
}
\arguments{
\item{pkg}{character string containing package name}
}
\value{
character vector of functions in given package
}
\description{
List functions in a package
}

