% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{lsos}
\alias{lsos}
\title{show largest objects}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{arguments passed on to \code{.ls.objects}}

\item{n}{scalar integer, number of objects to show}
}
\description{
https://gist.github.com/1187166.git Taken from
  http://stackoverflow.com/questions/1358003/\
  tricks-to-manage-the-available-memory-in-an-r-session
}
