% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{lsp}
\alias{lsp}
\title{List all items in a package}
\usage{
lsp(package, all.names = TRUE, pattern)
}
\arguments{
\item{package}{character scalar: name of the package}

\item{all.names}{= TRUE, set to FALSE to ignore items beginning with a period}

\item{pattern}{= optional pattern to match}
}
\value{
character vector of package contents
}
\description{
By default includes names beginning with '.'
}
\examples{
lsp("jwutil")
tail(lsp("base"), 30L)
}
