% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{mergeLists}
\alias{mergeLists}
\title{merge lists by names}
\usage{
mergeLists(x, y)
}
\arguments{
\item{x}{unnested list with named elements, each of which is a vector}

\item{y}{unnested list with named elements, each of which is a vector}
}
\value{
list of vectors
}
\description{
merge lists by vector combining all the vector elements of the
  list items with the matching names. Unnamed vectors in the list will be
  dropped silently.
}

