% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_percent.R
\name{npc}
\alias{npc}
\title{Print integers with percentage of total rounded to integer}
\usage{
npc(x, n, fmt = "\%d (\%s)")
}
\arguments{
\item{x}{numeric number}

\item{n}{numeric total}

\item{fmt}{`sprintf` format, default being `%%d (%%s)`}
}
\description{
Intended for succinctly printing summary data in a scientific publication.
}
\examples{
npc(1, 100)
npc(1, 1)
npc(2, 1)
npc(1.321, 7.7432)
npc(7239, 234897)
npc(-10, 1000)
}
