% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.r
\name{numbers_to_long_and_float}
\alias{numbers_to_long_and_float}
\title{convert numbers to long and float types}
\usage{
numbers_to_long_and_float(..., na.rm = TRUE)
}
\arguments{
\item{...}{list of values to convert to long and double}

\item{na.rm}{logical, defaults to TRUE, so output contains only long and
float values.}
}
\value{
list of long and double versions of convertable values from the input
}
\description{
intended for generating values for stress testing functions
}

