% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{opt_binary_brute}
\alias{opt_binary_brute}
\title{optimizes a function for all combinations of all subsets}
\usage{
opt_binary_brute(x, fun = opt_binary_fun, verbose = TRUE)
}
\arguments{
\item{x}{data frame}

\item{fun}{function which takes parameters x = data.frame, n = columns}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}
}
\description{
takes a data frame and optimization function
}
