% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.r
\docType{data}
\name{optName}
\alias{optName}
\title{option name for cache}
\format{\preformatted{ chr "cachedir"
}}
\usage{
optName
}
\description{
allow use of \code{options} to search for the cache directory.
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{assignCacheByFun}}, \code{\link{cache}};
  \code{\link{findCacheDir}};
  \code{\link{findCacheFilePath}};
  \code{\link{getCacheVarDated}}; \code{\link{isCached}};
  \code{\link{lsCache}}, \code{\link{lsCacheFiles}};
  \code{\link{renameCache}}; \code{\link{rmCache}};
  \code{\link{saveInDataDir}}; \code{\link{saveToCache}}
}
\keyword{character}
\keyword{internal}

