% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_percent.R
\name{percent_signif}
\alias{percent_signif}
\title{Return percentage string to given significant figures}
\usage{
percent_signif(x, figures = 3, sep = "")
}
\arguments{
\item{x}{numeric or integer values}

\item{figures}{integer number of significant figures to format}

\item{sep}{character used to separate number from percent symbol, default is
empty string}
}
\description{
From jwutil development version
}
