% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_percent.R
\name{percentize}
\alias{percentize}
\title{Convert a number into rounded integer percentage string}
\usage{
percentize(x)
}
\arguments{
\item{x}{numeric}
}
\description{
The number is converted into a percentage, then rounded.
}
\examples{
percentize(-1)
percentize(1L)
percentize(7.7)
percentize(0.01)
percentize(0.001)
}
