% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{permuteWithRepeats}
\alias{permuteWithRepeats}
\title{generate all permutations of input, reusing values in each result row}
\usage{
permuteWithRepeats(x)
}
\arguments{
\item{x}{list or vector}
}
\value{
data frame, each row being one permutation
}
\description{
systematically permute the input vector or list
}
