% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgo.R
\name{pgo_bench}
\alias{pgo_bench}
\title{Compile and test package with and without profile guided optimization Running
the test suite is a proxy for some kind of benchmark. Obviously this is
inadequate with most CRAN test-suites. However, the \code{test_that} path may
be set to a directory of \pkg{testthat} tests, which are called by
\code{\link[testthat]{testdir}}.}
\usage{
pgo_bench(pkg_path = getwd(), test_path = "tests/testthat",
  prof_dir = file.path(pkg_path, "src/prof"), iterations = 3,
  reporter = "silent", clean = FALSE, opt_level = "-O3",
  cc = "clang", cxx = "clang++", brew = Sys.info()[["sysname"]] ==
  "Darwin", ...)
}
\description{
Compile and test package with and without profile guided optimization Running
the test suite is a proxy for some kind of benchmark. Obviously this is
inadequate with most CRAN test-suites. However, the \code{test_that} path may
be set to a directory of \pkg{testthat} tests, which are called by
\code{\link[testthat]{testdir}}.
}
\keyword{internal}
