% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.r
\name{platformIsLinux}
\alias{platformIsLinux}
\alias{platformIsWindows}
\title{Are we running on Linux or Windows?}
\usage{
platformIsLinux()

platformIsWindows()
}
\value{
logical
}
\description{
Are we running on Linux or Windows?
}

