% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.r
\name{propIsNa}
\alias{propIsNa}
\title{Proportion of NA values in a vector}
\usage{
propIsNa(x)
}
\arguments{
\item{x}{is a vector which may have NA values}
}
\value{
numeric proportion of NAs in the supplied vector
}
\description{
Proportion of NA values in a vector
}

