% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/df.r
\name{propNaPerField}
\alias{propNaPerField}
\title{return proportion of NA values per field}
\usage{
propNaPerField(dframe)
}
\arguments{
\item{dframe}{is a data frame}
}
\value{
numeric vector
}
\description{
return proportion of NA values per field
}

