% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propRowSorted}
\alias{propRowSorted}
\title{proportion of non-descending rows in matrix}
\usage{
propRowSorted(x)
}
\arguments{
\item{x}{matrix, each row containing ordered or disordered numerics}
}
\value{
double, the proportion from 0 to 1
}
\description{
first performs isRowSorted to get a logical vector,
then sums TRUE values and takes fraction of total
}

