% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.r
\name{random_test_dates}
\alias{random_test_dates}
\alias{random_test_posixlt_datetimes}
\title{generate random Dates or POSIXlt test datetimes}
\usage{
random_test_dates(n = nums_in_tests, origin = as.Date("2000-01-01"),
  dayspread = 365 * 150)

random_test_posixlt_datetimes(n = nums_in_tests,
  origin = as.Date("2000-01-01"), dayspread = 365 * 150)
}
\arguments{
\item{n}{integer number to generate}

\item{origin}{Date defaults to Jan 1, 2000.}

\item{dayspread}{integer number of days either side of origin to pick random
dates from, defaults to 150 years.}
}
\value{
vector of POSIXlt datetimes or Dates
}
\description{
generate random \code{Date}s and \code{POSIXlt} test datetimes
}

