% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{read_xlsx_linux}
\alias{read_xlsx_linux}
\title{read \code{.xlsx} file, interpret as CSV, and return a data frame}
\usage{
read_xlsx_linux(file)
}
\arguments{
\item{file}{is the path to the \code{.xlsx} file}
}
\value{
data frame
}
\description{
currently relies on Linux xlsx2csv command, but could
  potentially be done with VB script in Windows. This offers a different
  backend to other Excel parsing functions in R,
}
\seealso{
\code{readxl} package by Hadley Wickham
}
