% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{readXlsxLinux}
\alias{readXlsxLinux}
\title{read .xlsx file, interpret as CSV, and return data frame}
\usage{
readXlsxLinux(file)
}
\arguments{
\item{file}{is the path to the .xlsx file}
}
\value{
data frame
}
\description{
currently relies on Linux xlsx2csv command, but could
  potentially be done with VB script in Windows
}

