% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{renameCache}
\alias{renameCache}
\title{rename cache file(s)}
\usage{
renameCache(oldname, newname, cacheDir = NULL, verbose = FALSE)
}
\arguments{
\item{oldname}{character vector length one}

\item{newname}{character vector length one}

\item{cacheDir}{character vector length one}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}
}
\description{
renames a cache file, or set of dated files. It has the ability
  to rename a single dated file, if the dated name is given, but probably
  best to keep all the files for the same root name in sync. The trick is not
  just renaming the files, but renaming the variable names of the data
  therein.
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{rmCache}}, \code{\link{saveInDataDir}},
  \code{\link{saveToCache}}
}

