% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{rmCache}
\alias{rmCache}
\title{remove from cache and given environment}
\usage{
rmCache(var = NULL, pattern = NULL, from = NULL, to = NULL,
  cacheDir = NULL)
}
\arguments{
\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{pattern}{character regular expression}

\item{from}{This should be a single YYYY-MM-DD string. There are no
absolute requirements about the format, but it is intended to be the same
format as used in a SQL query.}

\item{to}{see start date}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}
}
\description{
TOOD: warn if not found
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{saveInDataDir}},
  \code{\link{saveToCache}}
}

