% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rmRecursive}
\alias{rmRecursive}
\title{recursive remove}
\usage{
rmRecursive(x, envir = parent.frame())
}
\arguments{
\item{x}{variables to annihilate}

\item{envir}{environment to start at, defaults to calling frame.}
}
\description{
search through environments until the variables in the list
  \code{x} are all gone. This doesn't delete functions. No barrier to
  infinite recursion, but rm should be able to delete anything that exists
  can see.
}

