% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.r, R/util.r
\name{save_in_data_dir}
\alias{save_in_data_dir}
\title{Save given variable in package data directory}
\usage{
save_in_data_dir(var_name, suffix = "", package_dir = getwd(),
  envir = parent.frame())

save_in_data_dir(var_name, suffix = "", package_dir = getwd(),
  envir = parent.frame())
}
\arguments{
\item{var_name}{character or symbol, e.g. \code{"myvar"} or \code{myvar},
either of which would find \code{myvar} in the parent environment, and save
it as \code{myvar.RData} in \code{package_root/data}.}

\item{suffix}{character scalar}

\item{package_dir}{character containing the directory root of the package
tree in which to save the data. Default is the current working directory.}

\item{envir}{environment in which to look for the variable to save}

\item{data_path}{path to data directory, default is data in current directory.}

\item{var_name}{character or symbol, e.g. "myvar" or \code{myvar}, either of
which would find \code{myvar} in the parent environment, and save it as
\code{myvar.RData} in \code{package_root/data}.}

\item{suffix}{character scalar}
}
\value{
invisibly returns the data
}
\description{
File will be named \emph{var_name}\code{.RData} with an optional suffix
before \code{.RData}

File is named varname.RData with an optional suffix before .RData
}
\keyword{file}
\keyword{internal}

