% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{saveInDataDir}
\alias{saveInDataDir}
\title{save data compressed in data folder}
\usage{
saveInDataDir(var, suffix = "")
}
\arguments{
\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{suffix}{char additional characters before ".RData"}
}
\description{
xz appears to fail on Windows, so use bzip2

  tools::checkRdaFiles(file.path("data", list.files(path = "data")))
  tools::resaveRdaFiles(file.path("data", list.files(path = "data")),
  compress = "xz")
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveToCache}}
}

