% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{saveToCache}
\alias{saveToCache}
\title{Cache an R object}
\usage{
saveToCache(var, from = NULL, to = NULL, cacheDir = NULL,
  envir = parent.frame(), verbose = FALSE)
}
\arguments{
\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{from}{This should be a single YYYY-MM-DD string. There are no
absolute requirements about the format, but it is intended to be the same
format as used in a SQL query.}

\item{to}{see start date}

\item{cacheDir}{a single directory name of the cache to use. If not
specified, the default cache directory name will be sought}

\item{envir}{environment to start searching for the cached data (it may
already be in memory). Starts off with \code{parent.frame()} by default,
and /code{inherits}, so should find already loaded cache files in
.GlobalEnv eventually.}

\item{verbose}{single logical value, if \code{TRUE} then produce verbose
messages}
}
\value{
value in variable named by \code{var}
}
\description{
There are various memoise and cache functions in R but none did
  what I wanted. These functions allow a package to cache data in a standard
  place, or specified directory.

  Calling environment is used by default to save
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveInDataDir}}
}

