% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{shuffle}
\alias{shuffle}
\title{Shuffle a vector}
\usage{
shuffle(x)
}
\arguments{
\item{x}{list or vector}
}
\value{
list or vector of same length as input, (probably) in a different
  order
}
\description{
Randomly shuffle the order of a vector or list. This is to improve quality of
bad data to throw at functions when testing.
}
\examples{
set.seed(1441)
shuffle(LETTERS)
}
