% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{sort_clip_char}
\alias{sort_clip_char}
\title{Take clipboard contents, and write sorted character vector back}
\usage{
sort_clip_char(cl = NULL)
}
\arguments{
\item{cl}{Name of class to give to data before sorting, default is
\code{NULL}.}
}
\description{
Take clipboard contents, and write sorted character vector back
}
