% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{source_purl}
\alias{source_purl}
\title{extract code from knitr vignette and source it}
\usage{
source_purl(input, output = file.path(tempdir(), paste0(basename(input),
  ".R")), documentation = 1L, ...)
}
\arguments{
\item{input}{path to file as single character string}
}

