% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.r
\name{strip_for_formula}
\alias{strip_for_formula}
\title{strip a string so that it can be used as a variable name in a formula.}
\usage{
strip_for_formula(x)
}
\arguments{
\item{x}{character vector of potential formula variables}
}
\value{
character vector of length x
}
\description{
This excludes many symbols, so just strip all symbols leaving
  alphanumeric, and no whitespace.
}
