% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.r
\name{trim}
\alias{trim}
\title{strip whitespace from ends of each string in given character vector}
\usage{
trim(x)
}
\arguments{
\item{x}{is a character vector to trim}
}
\value{
character vector
}
\description{
slower than \code{strip}.
}
