% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github_pkgs.r
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{Update github_install packages}
\usage{
update_github_pkgs()
}
\value{
Returns invisibly the names of packages which need updating. THe
function outputs the commands to run to actually update them (by
reinstalling from github). Doesn't do this automatically because it would
mean bringing in a lot of dependencies.
}
\description{
Update github_install packages
}
