% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.r
\docType{data}
\name{zeroes}
\alias{zeroes}
\title{zeroes}
\format{\preformatted{List of 2
 $ : int 0
 $ : num 0
}}
\usage{
zeroes
}
\description{
long and float types
}
\keyword{sysdata}

