% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.r
\docType{data}
\name{bad_input}
\alias{bad_input}
\title{bad input data for tests}
\format{An object of class \code{list} of length 43.}
\usage{
bad_input
}
\description{
a variety of horrible data
}
\keyword{sysdata}
