% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{binaryCols}
\alias{binaryCols}
\alias{nonBinaryCols}
\alias{nonBinaryNumericCols}
\alias{non_binary_cols}
\alias{numericCols}
\title{names of fields which are numeric, binary or combinations thereof}
\usage{
binaryCols(x)

nonBinaryCols(x)

non_binary_cols(data, col_names)

numericCols(x)

nonBinaryNumericCols(x)
}
\arguments{
\item{x}{data frame}
}
\value{
vector of column names
}
\description{
Doesn't make any allowance for factors.
}
\examples{
dat <- data.frame(c("a", "b"), c(TRUE, FALSE), c(1, 0), c(1L, 0L),
                  c(1L, 2L), c(0.1, 0.2), c("9", "8"))
names(dat) <- c("char", "bin", "binfloat", "binint",
                "int", "float", "charint")
binaryCols(dat)
nonBinaryCols(dat)
numericCols(dat)
nonBinaryNumericCols(dat)
}

