% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{buildLinearFormula}
\alias{buildLinearFormula}
\title{build simple linear formula from variable names}
\usage{
buildLinearFormula(left, right)
}
\arguments{
\item{left}{character vector}

\item{right}{character vector}
}
\value{
formula
}

