% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.r
\docType{data}
\name{extreme_numbers}
\alias{extreme_numbers}
\title{extreme numbers}
\format{An object of class \code{numeric} of length 8.}
\usage{
extreme_numbers
}
\description{
very biggest and smallest non-zero numbers the current machine
  can handle, positive and negative.
}
\keyword{sysdata}

