% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{getCacheVarDated}
\alias{getCacheVarDated}
\title{make cache file name from dates and unadorned variable name}
\usage{
getCacheVarDated(var, from = NULL, to = NULL)
}
\arguments{
\item{var}{name of variable to assign. This may be used as a prefix
before a date-limited data extract, or as the entire variable name}

\item{from}{This should be a single YYYY-MM-DD string. There are no
absolute requirements about the format, but it is intended to be the same
format as used in a SQL query.}

\item{to}{see start date}
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{optName}},
  \code{\link{renameCache}}, \code{\link{rmCache}},
  \code{\link{saveInDataDir}}, \code{\link{saveToCache}}
}

