% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{getFactorNames}
\alias{getFactorNames}
\alias{getNonFactorNames}
\title{get names of the factor fields in a data frame}
\usage{
getFactorNames(x, consider = names(x))

getNonFactorNames(x, consider = names(x))
}
\arguments{
\item{x}{data frame}

\item{consider}{character vector of field names, default is to use all
of them.}
}
\value{
vector
}

