% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{getNAFields}
\alias{getNAFields}
\alias{getNonNAFields}
\title{get NA field names from data frame}
\usage{
getNAFields(dframe)

getNonNAFields(dframe)
}
\arguments{
\item{dframe}{data.frame}
}
\value{
vector of names of fields which contain any NA values, length zero if
  no matches
}

