% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwutil-package.r
\docType{package}
\name{jwutil-package}
\alias{jwutil}
\alias{jwutil-package}
\title{Tools for testing and data manipulation not found elsewhere}
\description{
This is a set of simple utilities for various data
    manipulation and testing tasks.  The goal is to use core R tools well,
    without bringing in many dependencies. Main areas of interest are
    semi-automated data frame manipulation, such as converting factors in
    multiple binary indicator columns. There are testing functions which
    provide 'testthat' expectations to permute arguments to function
    calls. There are functions and data to test extreme numbers, dates,
    and bad input of various kinds which should allow testing failure and
    corner cases, which can be used for fuzzing your functions. The test
    suite has many examples of usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jackwasey/jwutil}
  \item Report bugs at \url{https://github.com/jackwasey/jwutil/issues}
}

}
\author{
\strong{Maintainer}: Jack O. Wasey \email{jack@jackwasey.com} [copyright holder]

}
