% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwutil-package.r
\docType{package}
\name{jwutil-package}
\alias{jwutil}
\alias{jwutil-package}
\title{Tools for data manipulation not found elsewhere, and testthat extensions}
\description{
This is a set of simple utilities for various data manipulation and
caching tasks. The goal is to use base tools well, without bringing in any
dependencies. Main areas of interest are data frame manipulation, such as
converting factors in multiple binary indicator columns, and disk caching of
data frames (which is optionally done by date range). There are testing
functions which provide testthat expectations to permute arguments to
function calls. There are functions and data to test extreme numbers, dates,
and bad input of various kinds which should allow testing failure and corner
cases. The test suite has many examples of usage.
}

