% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{ls.objects}
\alias{ls.objects}
\title{Summarize objects in scope}
\usage{
ls.objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
\item{pos}{pos}

\item{pattern}{regex pattern to match objects of interest}

\item{order.by}{which column to order by}

\item{decreasing}{default is \code{TRUE}}

\item{head}{default is \code{FALSE} but if true, just show top \code{n}}

\item{n}{number to show if limiting to \code{head}}
}
\description{
Get type, size (bytes) and dimensions of objects in scope
}
