% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{lsp}
\alias{lsp}
\title{list all items in a package}
\usage{
lsp(package, all.names = TRUE, pattern)
}
\arguments{
\item{package}{is the (unquoted) name of the package}

\item{all.names}{= TRUE, set to FALSE to ignore items beginning with a period}

\item{pattern}{= optional pattern to match}
}
\value{
character vector of package contents
}
\description{
default to including (?private) functions beginning with '.'
}
