% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\docType{data}
\name{optName}
\alias{optName}
\title{option name for cache}
\format{An object of class \code{character} of length 1.}
\usage{
optName
}
\description{
allow use of \code{options} to search for the cache directory.
}
\seealso{
Other cache: \code{\link{assignCache}},
  \code{\link{findCacheDir}},
  \code{\link{findCacheFilePath}},
  \code{\link{getCacheVarDated}}, \code{\link{isCached}},
  \code{\link{lsCacheFiles}}, \code{\link{renameCache}},
  \code{\link{rmCache}}, \code{\link{saveInDataDir}},
  \code{\link{saveToCache}}
}
\keyword{character}
\keyword{internal}

