% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{permute}
\alias{permute}
\title{generate all permutations of input}
\usage{
permute(x)
}
\arguments{
\item{x}{list or vector}
}
\value{
data frame, each row being one permutation
}
\description{
systematically permute the input vector or list, which is very
  slow for long x. Am amazed something this simple isn't either in base R, or
  in a straightforward form in a package.

  TODO: limit to a certain cut-off, after which we randomly sample
}

