% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{chained functions and results}
}
\description{
Use the pipe function, \code{\%>\%} to turn function composition into a
series of imperative statements.
}
\keyword{internal}

