% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{propNaPerField}
\alias{propNaPerField}
\title{return proportion of NA values per field}
\usage{
propNaPerField(dframe)
}
\arguments{
\item{dframe}{is a data frame}
}
\value{
numeric vector
}
\description{
Return proportion of values which are \code{NA} in each field of
  the given data frame.
}
