% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{reqinst}
\alias{reqinst}
\title{install packages where are missing}
\usage{
reqinst(pkgs, quietly = TRUE, warn.conflicts = FALSE)
}
\arguments{
\item{pkgs}{vector of packages to load and install if necessary}

\item{quietly}{logical value, default \code{TRUE}}

\item{warn.conflicts}{logical value, default \code{FALSE}}
}
\description{
install packages where are missing
}
