% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{reqinst}
\alias{reqinst}
\title{install packages where are missing}
\usage{
reqinst(pkgs, ...)
}
\arguments{
\item{pkgs}{character vector of packages to load and attach, with
installation if necessary}

\item{...}{arguments passed to \code{library}}
}
\description{
install packages where are missing
}
