% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{reqinst}
\alias{reqinst}
\title{Load packages with \code{library}, installing any which are missing}
\usage{
reqinst(pkgs)
}
\arguments{
\item{pkgs}{character vector of packages to load and attach, with
installation if necessary}
}
\description{
Load packages with \code{library}, installing any which are missing
}
