% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github_pkgs.r
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{Update github_install packages}
\usage{
update_github_pkgs(no_action = FALSE)
}
\arguments{
\item{no_action}{Logical, if \code{TRUE}, no action will be taken, just a report
of what would have been done.}
}
\description{
Update github_install packages
}
