% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_na.r
\name{zero_na}
\alias{zero_na}
\title{Zero NA values in a data.frame}
\usage{
zero_na(df, cols = names(df), ignore = character(), verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{cols}{names of columns to work on, default is all columns}

\item{ignore}{character vector of columns names to ignore}

\item{verbose}{TRUE or FALSE}
}
\description{
Zero NA values in a data.frame, including \code{cols} and exluding
\code{ignore}. Also does not replace \code{Date} or \code{POSIXt} fields.
}
\examples{
d <- data.frame(1:5, 6:10, 11:15)
d[2, 3] <- NA
d[5, 2] <- NA
d[1, 1] <- NA
print(d)
zero_na(d)
}
