% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.r
\name{zero_na}
\alias{zero_na}
\title{zero NA values in a data.frame}
\usage{
zero_na(df, cols = names(df), ignore = character(), verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{cols}{names of columns to work on, default is all columns}

\item{ignore}{cahracter vector of columns names to ignore}

\item{verbose}{TRUE or FALSE}
}
\description{
Zero NA values in a data.frame, including \code{cols} and
  exluding \code{ignore}. Also does not replace \code{Date} or \code{POSIXt}
  fields.
}

